add missing dependency tests to CI
