+++ .github/workflows/test.yml
@@ -91,11 +91,14 @@ jobs:
       run: |
         poetry config virtualenvs.in-project true
         poetry run pip install --upgrade pip
-        poetry install -E all
+        poetry install
         poetry run pip install -U "Django~=${{ matrix.django-version }}"
     - name: Run Unit Tests
       run: |
-        poetry run pytest
+        pip uninstall -y Jinja2 PyYAML
+        poetry run pytest --cov-append
+        poetry install -E all
+        poetry run pytest --cov-append
         mv .coverage py${{ matrix.python-version }}-dj${{ matrix.django-version }}.coverage
 
     - name: Store coverage files
